Commit 1 - file struture
