+++ server.js
@@ -1,6 +1,7 @@
 // DEPENDENCIES
 const express = require('express');
 const app = express();
+const marsMissions = require('./models/marsMissions');
 
 // run `npm install` to install dependencies in package.json
 
@@ -22,43 +23,7 @@ const port = 3000;
 
 // DATA - put into marsMissions.js file inside of a models folder, for module.exports
 // remember to require it in the server
-const marsMissions = [
-  {
-    name: "Curiosity",
-    launchDate: "26 Nov 2011",
-    operator: "NASA",
-    missionType: "Rover",
-    img: ""
-  },
-  {
-    name: "Opportunity",
-    launchDate: "8 Jul 2003",
-    operator: "NASA",
-    missionType: "Rover",
-    img: ""
-  },
-  {
-    name: "Spirit",
-    launchDate: "10 Jun 2003",
-    operator: "NASA",
-    missionType: "Rover",
-    img: ""
-  },
-  {
-    name: "Sojourner",
-    launchDate: "4 Dec 1996",
-    operator: "NASA",
-    missionType: "Rover",
-    img: ""
-  },
-  {
-    name: "Rosetta",
-    launchDate: "2 Mar 2004",
-    operator: "ESA",
-    missionType: "Gravity Assist",
-    img: ""
-  }
-];
+
 
 // INDEX Route
 // send data to 'missions/index.ejs' view
@@ -71,9 +36,17 @@ const marsMissions = [
 
 
 
+
+
+module.exports = app;
+
+app.get('/missions', (req, res) => {
+  res.render('missions/show.ejs', {
+    "marsMissions": marsMissions
+  })
+})
+
 // LISTENER
 app.listen(port, function() {
   console.log('Missions to Mars running on port: ', port);
 })
-
-module.exports = app;
